+++ app/modules/depositor_admin_setting/depositor_admin_setting/__init__.py
@@ -1,0 +1,3 @@
+
+from .ext import AdminSettingApp
+__all__ = ('AdminSettingApp')
