+++ app/modules/depositor_admin_setting/depositor_admin_setting/static/js/admin_settings.js
