+++ app/modules/depositor_grobid_client/setup.py
@@ -1,12 +1,12 @@
 from setuptools import setup, find_packages
 
 setup(
-    name='login',
+    name='depositor_grobid_client',
     version='1.0.0',
     author='Your Name',
     author_email='your.email@example.com',
     description='A short description of your package',
-    packages=["login"],
+    packages=find_packages(),
     install_requires=[
     ],
     entry_points={
