+++ app/modules/depositor_item_register/setup.py
@@ -1,0 +1,14 @@
+from setuptools import setup, find_packages
+
+setup(
+    name='depositor_item_register',
+    version='1.0.0',
+    author='Your Name',
+    author_email='your.email@example.com',
+    description='A short description of your package',
+    packages=find_packages(),
+    install_requires=[
+    ],
+    entry_points={
+    },
+)
