+++ app/modules/depositor_login/depositor_login/__init__.py
@@ -1,0 +1,3 @@
+
+from .ext import LoginApp
+__all__ = ('LoginApp')
