+++ app/modules/depositor_login/depositor_login/views.py
@@ -7,9 +7,8 @@ from flask_security import LoginForm, url_for_security
 from flask_wtf import FlaskForm
 from wtforms import StringField, SubmitField ,PasswordField
 from .config import MOCK_SHIB_DATA
-from modules.models.models.models import User as _User
-from modules.models.models.api import User
-from modules.models.models.api import Affiliation_Id
+from depositor_models.user import User, User_manager
+from depositor_models.affiliation_id import Affiliation_Id_manager
 
 blueprint = Blueprint(
     "login",
@@ -29,6 +28,9 @@ def top():
 
 @blueprint.route("/login", methods=['GET'])
 def index_login():
+    import sys
+
+    print(sys.path)
     if not current_user.is_anonymous:
         return redirect(url_for('item_register.index_item'))
     csrf_random = generate_random_str(length=64)
@@ -46,24 +48,24 @@ def login():
             shib_data = MOCK_SHIB_DATA.get(form.email.data)
             if shib_data :
                 affiliation_idp_url = shib_data.get("affiliation_idp_url",None)
-                affiliation_id = Affiliation_Id().get_affiliation_id_by_idp_url(affiliation_idp_url)
+                affiliation_id = Affiliation_Id_manager().get_affiliation_id_by_idp_url(affiliation_idp_url)
                 if not affiliation_id:
                     affiliation_name = shib_data.get("OrganizationName",None)
-                    affiliation_id = Affiliation_Id().create_affiliation_id(affiliation_idp_url=affiliation_idp_url,
+                    affiliation_id = Affiliation_Id_manager().create_affiliation_id(affiliation_idp_url=affiliation_idp_url,
                                                                       affiliation_name=affiliation_name)
                 affiliation_id_id = affiliation_id.id
                 user_id = shib_data.get("eduPersonPrincipalName",None)
-                user=User().get_user_by_user_id(user_id)
+                user=User_manager().get_user_by_user_id(user_id)
                 if not user :
                     user_orcid = shib_data.get("eduPersonOrcid",None)
                     role = shib_data.get("wekoSocietyAffiliation", None)
-                    user = _User(
+                    user = User(
                         user_id = user_id,
                         affiliation_id = affiliation_id_id,
                         user_orcid = user_orcid,
                         role = role
                     )
-                    User().create_user(user)
+                    User_manager().create_user(user)
                 login_user(user)
                 return redirect(url_for("item_register.index_item"))
     flash("Missing SHIB_ATTRs!", category='error')
