+++ app/modules/depositor_login/tests/test_views.py
@@ -1,4 +1,4 @@
-from login.views import generate_random_str
+from depositor_login.views import generate_random_str
 
 def test_generate_random_str():
     assert generate_random_str()
