--- app/modules/login/coverage.xml
@@ -1,103 +1,0 @@
-<?xml version="1.0" ?>
-<coverage version="7.4.3" timestamp="1709192671534" lines-valid="70" lines-covered="0" line-rate="0" branches-covered="0" branches-valid="0" branch-rate="0" complexity="0">
-	<!-- Generated by coverage.py: https://coverage.readthedocs.io/en/7.4.3 -->
-	<!-- Based on https://raw.githubusercontent.com/cobertura/web/master/htdocs/xml/coverage-04.dtd -->
-	<sources>
-		<source>/code/modules/login/login</source>
-	</sources>
-	<packages>
-		<package name="." line-rate="0" branch-rate="0" complexity="0">
-			<classes>
-				<class name="__ini__.py" filename="__ini__.py" complexity="0" line-rate="1" branch-rate="0">
-					<methods/>
-					<lines/>
-				</class>
-				<class name="config.py" filename="config.py" complexity="0" line-rate="0" branch-rate="0">
-					<methods/>
-					<lines>
-						<line number="1" hits="0"/>
-						<line number="3" hits="0"/>
-						<line number="6" hits="0"/>
-						<line number="8" hits="0"/>
-					</lines>
-				</class>
-				<class name="ext.py" filename="ext.py" complexity="0" line-rate="0" branch-rate="0">
-					<methods/>
-					<lines>
-						<line number="1" hits="0"/>
-						<line number="3" hits="0"/>
-						<line number="4" hits="0"/>
-						<line number="5" hits="0"/>
-						<line number="6" hits="0"/>
-						<line number="8" hits="0"/>
-						<line number="9" hits="0"/>
-						<line number="10" hits="0"/>
-					</lines>
-				</class>
-				<class name="views.py" filename="views.py" complexity="0" line-rate="0" branch-rate="0">
-					<methods/>
-					<lines>
-						<line number="2" hits="0"/>
-						<line number="3" hits="0"/>
-						<line number="4" hits="0"/>
-						<line number="5" hits="0"/>
-						<line number="6" hits="0"/>
-						<line number="7" hits="0"/>
-						<line number="8" hits="0"/>
-						<line number="9" hits="0"/>
-						<line number="10" hits="0"/>
-						<line number="11" hits="0"/>
-						<line number="12" hits="0"/>
-						<line number="14" hits="0"/>
-						<line number="21" hits="0"/>
-						<line number="22" hits="0"/>
-						<line number="23" hits="0"/>
-						<line number="24" hits="0"/>
-						<line number="26" hits="0"/>
-						<line number="27" hits="0"/>
-						<line number="28" hits="0"/>
-						<line number="30" hits="0"/>
-						<line number="31" hits="0"/>
-						<line number="32" hits="0"/>
-						<line number="33" hits="0"/>
-						<line number="34" hits="0"/>
-						<line number="35" hits="0"/>
-						<line number="36" hits="0"/>
-						<line number="38" hits="0"/>
-						<line number="40" hits="0"/>
-						<line number="41" hits="0"/>
-						<line number="42" hits="0"/>
-						<line number="43" hits="0"/>
-						<line number="45" hits="0"/>
-						<line number="46" hits="0"/>
-						<line number="47" hits="0"/>
-						<line number="48" hits="0"/>
-						<line number="49" hits="0"/>
-						<line number="50" hits="0"/>
-						<line number="51" hits="0"/>
-						<line number="52" hits="0"/>
-						<line number="54" hits="0"/>
-						<line number="55" hits="0"/>
-						<line number="56" hits="0"/>
-						<line number="57" hits="0"/>
-						<line number="58" hits="0"/>
-						<line number="59" hits="0"/>
-						<line number="60" hits="0"/>
-						<line number="66" hits="0"/>
-						<line number="67" hits="0"/>
-						<line number="68" hits="0"/>
-						<line number="69" hits="0"/>
-						<line number="70" hits="0"/>
-						<line number="73" hits="0"/>
-						<line number="74" hits="0"/>
-						<line number="75" hits="0"/>
-						<line number="76" hits="0"/>
-						<line number="79" hits="0"/>
-						<line number="81" hits="0"/>
-						<line number="83" hits="0"/>
-					</lines>
-				</class>
-			</classes>
-		</package>
-	</packages>
-</coverage>
