+++ app/requirements-depositor-modules.txt
@@ -1,0 +1,5 @@
+/code/modules/depositor_admin_setting
+/code/modules/depositor_grobid_client
+/code/modules/depositor_item_register
+/code/modules/depositor_login
+/code/modules/depositor_models
